refactor/rename round -> trial
